--- src/backend/controllers/search.ts
@@ -1,19 +1,0 @@
-import { RequestHandler } from 'express'
-import { ENV } from '../enums/environments'
-import config from '../config'
-
-class SearchController {
-  public search: RequestHandler = (req, res, next) => {
-    const isIntegrationEnv = config.environment === ENV.DEVELOPMENT
-
-    try {
-      res.render('search.njk', {
-        isIntegrationEnv,
-      })
-    } catch (e) {
-      next(e)
-    }
-  }
-}
-
-export default SearchController
