+++ src/backend/routes/IndexRoutes.ts
@@ -1,8 +1,8 @@
 import { Router } from 'express'
-import SearchController from '../controllers/search'
-import SearchAPIController from '../controllers/searchAPI'
-import InfoBoxController from '../controllers/infoBox'
-import DownloadCSVController from '../controllers/downloadCSV'
+import SearchController from '../controllers/search.controller';
+import SearchAPIController from '../controllers/searchAPI.controller';
+import InfoBoxController from '../controllers/infoBox.controller';
+import DownloadCSVController from '../controllers/downloadCSV.controller';
 import Routes, { Route } from '../enums/routes'
 import { auth } from '../middleware/auth'
